requirements.txt
blinker==1.6.3
click==8.1.7
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
plotly==5.18.0
tenacity==8.2.3
typing_extensions==4.8.0
Werkzeug==3.0.1
